+++ src/main/java/com/tagmycode/plugin/gui/field/AbstractFieldValidation.java
@@ -19,7 +19,7 @@ public abstract class AbstractFieldValidation {
 
     public boolean performValidation() {
         if (!validate()) {
-            framework.showError(getMessageError());
+            framework.showErrorDialog(getMessageError());
             field.requestFocus();
             return false;
         }